more reorg, org mode print tests
